Fixed pep8 compliance; moved multi_mol.py to FOX.classes
