+++ FOX/__init__.py
@@ -8,8 +8,9 @@ from .__version__ import __version__
 from .functions import (
     read_multi_xyz,
     get_rdf, get_rdf_lowmem,
-    MultiMolecule
 )
+
+from .classes import MultiMolecule
 
 __all__ = [
     'read_multi_xyz',
