+++ FOX/classes/__init__.py
@@ -1,0 +1,5 @@
+from .multi_mol import MultiMolecule
+
+__all__ = [
+    'MultiMolecule'
+]
