+++ FOX/classes/multi_mol_magic.py
