+++ FOX/functions/__init__.py
@@ -1,9 +1,7 @@
-from .read_xyz import (read_multi_xyz)
 from .rdf import (get_rdf_lowmem, get_rdf)
-from .multi_mol import MultiMolecule
+from .read_xyz import read_multi_xyz
 
 __all__ = [
-    'read_multi_xyz',
     'get_rdf_lowmem', 'get_rdf',
-    'MultiMolecule'
-    ]
+    'read_multi_xyz'
+]
