+++ setup.py
@@ -26,7 +26,8 @@ setup(
         'FOX',
         'FOX.data',
         'FOX.examples',
-        'FOX.functions'
+        'FOX.functions',
+        'FOX.classes'
     ],
     package_dir={'FOX': 'FOX'},
     package_data={'FOX': ['data/Cd68Se55_26COO_MD_trajec.xyz']},
